Move ConnectionGroup
It was inside a folder called `network/`, called `ConnectionGroup.js`
and it was throwing off the vibe of the whole project. Now it's in a
dedicated folder named `connection-group`. Yay!
